test: fix resiliency race condition
